리팩토링: module.css 추가 및 Board 분할
