+++ src/Board.module.css
@@ -25,7 +25,7 @@
   position: absolute;
   top: 60%;
   left: 50%;
-  transform: translate(-50%, -50%);
+  transform: translate(-50%, 0%);
   background-color: rgba(238, 228, 218, 0.8);
   padding: 20px;
   font-size: 32px;
